Clean up metricsprocessor
