+++ pkg/otelcollector/metricsprocessor/processor.go
@@ -3,8 +3,6 @@ package metricsprocessor
 import (
 	"context"
 	"fmt"
-	"strings"
-	"time"
 
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/consumer"
@@ -15,6 +13,7 @@ import (
 	"github.com/fluxninja/aperture/pkg/log"
 	"github.com/fluxninja/aperture/pkg/metrics"
 	"github.com/fluxninja/aperture/pkg/otelcollector"
+	"github.com/fluxninja/aperture/pkg/otelcollector/metricsprocessor/internal"
 	"github.com/fluxninja/aperture/pkg/policies/dataplane/iface"
 	"github.com/rs/zerolog"
 )
@@ -75,238 +74,37 @@ func (p *metricsProcessor) ConsumeLogs(ctx context.Context, ld plog.Logs) (plog.
 				return retErr("aperture check response label not found in Envoy access logs")
 			}
 
-			addSDKSpecificLabels(attributes)
+			internal.AddSDKSpecificLabels(attributes)
 		} else if sourceStr == otelcollector.ApertureSourceEnvoy {
 			success := otelcollector.GetStruct(attributes, otelcollector.ApertureCheckResponseLabel, checkResponse, []string{otelcollector.EnvoyMissingAttributeValue})
 			if !success {
 				return retErr("aperture check response label not found in SDK access logs")
 			}
 
-			addEnvoySpecificLabels(attributes)
+			internal.AddEnvoySpecificLabels(attributes)
 		} else {
 			return retErr("aperture source label not recognized")
 		}
 
-		statusCode, featureStatus := statusesFromAttributes(attributes)
-		attributes.PutStr(otelcollector.ApertureResponseStatusLabel, responseStatusForTelemetry(statusCode, featureStatus))
-		addCheckResponseBasedLabels(attributes, checkResponse, sourceStr)
+		statusCode, featureStatus := internal.StatusesFromAttributes(attributes)
+		attributes.PutStr(otelcollector.ApertureResponseStatusLabel, internal.ResponseStatusForTelemetry(statusCode, featureStatus))
+		internal.AddCheckResponseBasedLabels(attributes, checkResponse, sourceStr)
 
 		// Update metrics and enforce include list to eliminate any excess attributes
 		if sourceStr == otelcollector.ApertureSourceSDK {
 			p.updateMetrics(attributes, checkResponse, []string{})
-			enforceIncludeListSDK(attributes)
+			internal.EnforceIncludeListSDK(attributes)
 		} else if sourceStr == otelcollector.ApertureSourceEnvoy {
 			p.updateMetrics(attributes, checkResponse, []string{otelcollector.EnvoyMissingAttributeValue})
-			enforceIncludeListHTTP(attributes)
+			internal.EnforceIncludeListHTTP(attributes)
 		}
 
 		// Add dynamic Flow labels
-		addFlowLabels(attributes, checkResponse)
+		internal.AddFlowLabels(attributes, checkResponse)
 
 		return nil
 	})
 	return ld, err
-}
-
-func addSDKSpecificLabels(attributes pcommon.Map) {
-	// Compute durations
-	flowStart, flowStartExists := getSDKLabelTimestampValue(attributes, otelcollector.ApertureFlowStartTimestampLabel)
-	workloadStart, workloadStartExists := getSDKLabelTimestampValue(attributes, otelcollector.ApertureWorkloadStartTimestampLabel)
-	flowEnd, flowEndExists := getSDKLabelTimestampValue(attributes, otelcollector.ApertureFlowEndTimestampLabel)
-
-	if flowStartExists && flowEndExists {
-		flowDuration := flowEnd.Sub(flowStart)
-		attributes.PutDouble(otelcollector.FlowDurationLabel, float64(flowDuration.Milliseconds()))
-	}
-	if workloadStartExists && flowEndExists {
-		workloadDuration := flowEnd.Sub(workloadStart)
-		attributes.PutDouble(otelcollector.WorkloadDurationLabel, float64(workloadDuration.Milliseconds()))
-	}
-}
-
-func addEnvoySpecificLabels(attributes pcommon.Map) {
-	treatAsZero := []string{otelcollector.EnvoyMissingAttributeValue}
-	// Retrieve request length
-	requestLength, _ := otelcollector.GetFloat64(attributes, otelcollector.EnvoyBytesSentLabel, treatAsZero)
-	attributes.PutDouble(otelcollector.HTTPRequestContentLength, requestLength)
-	// Retrieve response lengths
-	responseLength, _ := otelcollector.GetFloat64(attributes, otelcollector.EnvoyBytesReceivedLabel, treatAsZero)
-	attributes.PutDouble(otelcollector.HTTPResponseContentLength, responseLength)
-
-	// Compute durations
-	responseDuration, responseDurationExists := otelcollector.GetFloat64(attributes, otelcollector.EnvoyResponseDurationLabel, treatAsZero)
-	authzDuration, authzDurationExists := otelcollector.GetFloat64(attributes, otelcollector.EnvoyAuthzDurationLabel, treatAsZero)
-
-	if responseDurationExists {
-		attributes.PutDouble(otelcollector.FlowDurationLabel, responseDuration)
-	}
-
-	if responseDurationExists && authzDurationExists {
-		attributes.PutDouble(otelcollector.WorkloadDurationLabel, responseDuration-authzDuration)
-	}
-}
-
-func getSDKLabelTimestampValue(attributes pcommon.Map, labelKey string) (time.Time, bool) {
-	return getLabelTimestampValue(attributes, labelKey, otelcollector.ApertureSourceSDK)
-}
-
-func getLabelTimestampValue(attributes pcommon.Map, labelKey, source string) (time.Time, bool) {
-	value, exists := getLabelValue(attributes, labelKey, source)
-	if !exists {
-		return time.Time{}, false
-	}
-	return _getLabelTimestampValue(value, labelKey, source)
-}
-
-func _getLabelTimestampValue(value pcommon.Value, labelKey, source string) (time.Time, bool) {
-	var valueInt int64
-	if value.Type() == pcommon.ValueTypeInt {
-		valueInt = value.Int()
-	} else {
-		log.Sample(zerolog.Sometimes).Warn().Str("source", source).Str("key", labelKey).Msg("Failed to parse a timestamp field")
-		return time.Time{}, false
-	}
-
-	return time.Unix(0, valueInt), true
-}
-
-func getLabelValue(attributes pcommon.Map, labelKey, source string) (pcommon.Value, bool) {
-	value, exists := attributes.Get(labelKey)
-	if !exists {
-		log.Sample(zerolog.Sometimes).Warn().Str("source", source).Str("key", labelKey).Msg("Label not found")
-		return pcommon.Value{}, false
-	}
-	return value, exists
-}
-
-// addCheckResponseBasedLabels adds the following labels:
-// * otelcollector.ApertureProcessingDurationLabel
-// * otelcollector.ApertureServicesLabel
-// * otelcollector.ApertureControlPointLabel
-// * otelcollector.ApertureRateLimitersLabel
-// * otelcollector.ApertureDroppingRateLimitersLabel
-// * otelcollector.ApertureConcurrencyLimitersLabel
-// * otelcollector.ApertureDroppingConcurrencyLimitersLabel
-// * otelcollector.ApertureWorkloadsLabel
-// * otelcollector.ApertureDroppingWorkloadsLabel
-// * otelcollector.ApertureFluxMetersLabel
-// * otelcollector.ApertureFlowLabelKeysLabel
-// * otelcollector.ApertureClassifiersLabel
-// * otelcollector.ApertureClassifierErrorsLabel
-// * otelcollector.ApertureDecisionTypeLabel
-// * otelcollector.ApertureRejectReasonLabel
-// * otelcollector.ApertureErrorLabel.
-func addCheckResponseBasedLabels(attributes pcommon.Map, checkResponse *flowcontrolv1.CheckResponse, sourceStr string) {
-	// Aperture Processing Duration
-	startTime := checkResponse.GetStart().AsTime()
-	endTime := checkResponse.GetEnd().AsTime()
-	if !startTime.IsZero() && !endTime.IsZero() {
-		attributes.PutDouble(otelcollector.ApertureProcessingDurationLabel, float64(endTime.Sub(startTime).Milliseconds()))
-	} else {
-		log.Sample(zerolog.Sometimes).Warn().Msgf("Aperture processing duration not found in %s access logs", sourceStr)
-	}
-	// Services
-	servicesValue := pcommon.NewValueSlice()
-	for _, service := range checkResponse.Services {
-		servicesValue.Slice().AppendEmpty().SetStr(service)
-	}
-	servicesValue.CopyTo(attributes.PutEmpty(otelcollector.ApertureServicesLabel))
-
-	// Control Point
-	attributes.PutString(otelcollector.ApertureControlPointLabel, checkResponse.GetControlPointInfo().String())
-
-	labels := map[string]pcommon.Value{
-		otelcollector.ApertureRateLimitersLabel:                pcommon.NewValueSlice(),
-		otelcollector.ApertureDroppingRateLimitersLabel:        pcommon.NewValueSlice(),
-		otelcollector.ApertureConcurrencyLimitersLabel:         pcommon.NewValueSlice(),
-		otelcollector.ApertureDroppingConcurrencyLimitersLabel: pcommon.NewValueSlice(),
-		otelcollector.ApertureWorkloadsLabel:                   pcommon.NewValueSlice(),
-		otelcollector.ApertureDroppingWorkloadsLabel:           pcommon.NewValueSlice(),
-		otelcollector.ApertureFluxMetersLabel:                  pcommon.NewValueSlice(),
-		otelcollector.ApertureFlowLabelKeysLabel:               pcommon.NewValueSlice(),
-		otelcollector.ApertureClassifiersLabel:                 pcommon.NewValueSlice(),
-		otelcollector.ApertureClassifierErrorsLabel:            pcommon.NewValueSlice(),
-		otelcollector.ApertureDecisionTypeLabel:                pcommon.NewValueStr(checkResponse.DecisionType.String()),
-		otelcollector.ApertureRejectReasonLabel:                pcommon.NewValueStr(checkResponse.GetRejectReason().String()),
-		otelcollector.ApertureErrorLabel:                       pcommon.NewValueStr(checkResponse.GetError().String()),
-	}
-	for _, decision := range checkResponse.LimiterDecisions {
-		if decision.GetRateLimiterInfo() != nil {
-			rawValue := []string{
-				fmt.Sprintf("%s:%v", metrics.PolicyNameLabel, decision.GetPolicyName()),
-				fmt.Sprintf("%s:%v", metrics.ComponentIndexLabel, decision.GetComponentIndex()),
-				fmt.Sprintf("%s:%v", metrics.PolicyHashLabel, decision.GetPolicyHash()),
-			}
-			value := strings.Join(rawValue, ",")
-			labels[otelcollector.ApertureRateLimitersLabel].Slice().AppendEmpty().SetStr(value)
-			if decision.Dropped {
-				labels[otelcollector.ApertureDroppingRateLimitersLabel].Slice().AppendEmpty().SetStr(value)
-			}
-		}
-		if cl := decision.GetConcurrencyLimiterInfo(); cl != nil {
-			rawValue := []string{
-				fmt.Sprintf("%s:%v", metrics.PolicyNameLabel, decision.GetPolicyName()),
-				fmt.Sprintf("%s:%v", metrics.ComponentIndexLabel, decision.GetComponentIndex()),
-				fmt.Sprintf("%s:%v", metrics.PolicyHashLabel, decision.GetPolicyHash()),
-			}
-			value := strings.Join(rawValue, ",")
-			labels[otelcollector.ApertureConcurrencyLimitersLabel].Slice().AppendEmpty().SetStr(value)
-			if decision.Dropped {
-				labels[otelcollector.ApertureDroppingConcurrencyLimitersLabel].Slice().AppendEmpty().SetStr(value)
-			}
-
-			workloadsRawValue := []string{
-				fmt.Sprintf("%s:%v", metrics.PolicyNameLabel, decision.GetPolicyName()),
-				fmt.Sprintf("%s:%v", metrics.ComponentIndexLabel, decision.GetComponentIndex()),
-				fmt.Sprintf("%s:%v", metrics.WorkloadIndexLabel, cl.GetWorkloadIndex()),
-				fmt.Sprintf("%s:%v", metrics.PolicyHashLabel, decision.GetPolicyHash()),
-			}
-			value = strings.Join(workloadsRawValue, ",")
-			labels[otelcollector.ApertureWorkloadsLabel].Slice().AppendEmpty().SetStr(value)
-			if decision.Dropped {
-				labels[otelcollector.ApertureDroppingWorkloadsLabel].Slice().AppendEmpty().SetStr(value)
-			}
-		}
-	}
-	for _, fluxMeter := range checkResponse.FluxMeterInfos {
-		value := fluxMeter.GetFluxMeterName()
-		labels[otelcollector.ApertureFluxMetersLabel].Slice().AppendEmpty().SetStr(value)
-	}
-
-	for _, flowLabelKey := range checkResponse.GetFlowLabelKeys() {
-		labels[otelcollector.ApertureFlowLabelKeysLabel].Slice().AppendEmpty().SetStr(flowLabelKey)
-	}
-
-	for _, classifier := range checkResponse.ClassifierInfos {
-		rawValue := []string{
-			fmt.Sprintf("%s:%v", metrics.PolicyNameLabel, classifier.PolicyName),
-			fmt.Sprintf("%s:%v", metrics.ClassifierIndexLabel, classifier.ClassifierIndex),
-		}
-		value := strings.Join(rawValue, ",")
-		labels[otelcollector.ApertureClassifiersLabel].Slice().AppendEmpty().SetStr(value)
-
-		// add errors as attributes as well
-		if classifier.Error != flowcontrolv1.ClassifierInfo_ERROR_NONE {
-			errorsValue := []string{
-				classifier.Error.String(),
-				fmt.Sprintf("%s:%v", metrics.PolicyNameLabel, classifier.PolicyName),
-				fmt.Sprintf("%s:%v", metrics.ClassifierIndexLabel, classifier.ClassifierIndex),
-				fmt.Sprintf("%s:%v", metrics.PolicyHashLabel, classifier.PolicyHash),
-			}
-			joinedValue := strings.Join(errorsValue, ",")
-			labels[otelcollector.ApertureClassifierErrorsLabel].Slice().AppendEmpty().SetStr(joinedValue)
-		}
-	}
-
-	for key, value := range labels {
-		value.CopyTo(attributes.PutEmpty(key))
-	}
-}
-
-func addFlowLabels(attributes pcommon.Map, checkResponse *flowcontrolv1.CheckResponse) {
-	for key, value := range checkResponse.TelemetryFlowLabels {
-		pcommon.NewValueStr(value).CopyTo(attributes.PutEmpty(key))
-	}
 }
 
 func (p *metricsProcessor) updateMetrics(
@@ -348,7 +146,7 @@ func (p *metricsProcessor) updateMetrics(
 
 	if len(checkResponse.FluxMeterInfos) > 0 {
 		// Update flux meter metrics
-		statusCode, featureStatus := statusesFromAttributes(attributes)
+		statusCode, featureStatus := internal.StatusesFromAttributes(attributes)
 		for _, fluxMeter := range checkResponse.FluxMeterInfos {
 			p.updateMetricsForFluxMeters(
 				fluxMeter,
@@ -443,109 +241,9 @@ func (p *metricsProcessor) updateMetricsForFluxMeters(
 	// metricValue is the value at fluxMeter's AttributeKey
 	metricValue, _ := otelcollector.GetFloat64(attributes, fluxMeter.GetAttributeKey(), treatAsZero)
 
-	labels := statusLabelsForMetrics(decisionType, statusCode, featureStatus)
+	labels := internal.StatusLabelsForMetrics(decisionType, statusCode, featureStatus)
 	fluxMeterHistogram := fluxMeter.GetHistogram(labels)
 	if fluxMeterHistogram != nil {
 		fluxMeterHistogram.Observe(metricValue)
 	}
-}
-
-func statusesFromAttributes(attributes pcommon.Map) (statusCode string, featureStatus string) {
-	rawStatusCode, exists := attributes.Get(otelcollector.HTTPStatusCodeLabel)
-	if exists {
-		statusCode = rawStatusCode.Str()
-	}
-	rawFeatureStatus, exists := attributes.Get(otelcollector.ApertureFeatureStatusLabel)
-	if exists {
-		featureStatus = rawFeatureStatus.Str()
-	}
-	return
-}
-
-func statusLabelsForMetrics(
-	decisionType flowcontrolv1.CheckResponse_DecisionType,
-	statusCode string,
-	featureStatus string,
-) map[string]string {
-	return map[string]string{
-		metrics.ResponseStatusLabel: responseStatusForMetrics(statusCode, featureStatus),
-		metrics.DecisionTypeLabel:   decisionType.String(),
-		metrics.StatusCodeLabel:     statusCode,
-		metrics.FeatureStatusLabel:  featureStatus,
-	}
-}
-
-func responseStatusForMetrics(statusCode, featureStatus string) string {
-	return responseStatus(
-		statusCode,
-		featureStatus,
-		metrics.ResponseStatusOK,
-		metrics.ResponseStatusError)
-}
-
-func responseStatusForTelemetry(statusCode, featureStatus string) string {
-	return responseStatus(
-		statusCode,
-		featureStatus,
-		otelcollector.ApertureResponseStatusOK,
-		otelcollector.ApertureResponseStatusError)
-}
-
-func responseStatus(statusCode, featureStatus, okStatus, errorStatus string) string {
-	if strings.HasPrefix(statusCode, "2") {
-		return okStatus
-	}
-	if featureStatus != "" {
-		return featureStatus
-	}
-	return errorStatus
-}
-
-/*
- * IncludeList: This IncludeList is applied to logs and spans at during the enrichment process, after check response based labels are attached and metrics have been parsed.
- */
-var (
-	_includeAttributesCommon = []string{
-		otelcollector.ApertureSourceLabel,
-		otelcollector.WorkloadDurationLabel,
-		otelcollector.FlowDurationLabel,
-		otelcollector.ApertureProcessingDurationLabel,
-		otelcollector.ApertureDecisionTypeLabel,
-		otelcollector.ApertureErrorLabel,
-		otelcollector.ApertureRejectReasonLabel,
-		otelcollector.ApertureRateLimitersLabel,
-		otelcollector.ApertureDroppingRateLimitersLabel,
-		otelcollector.ApertureConcurrencyLimitersLabel,
-		otelcollector.ApertureDroppingConcurrencyLimitersLabel,
-		otelcollector.ApertureWorkloadsLabel,
-		otelcollector.ApertureDroppingWorkloadsLabel,
-		otelcollector.ApertureFluxMetersLabel,
-		otelcollector.ApertureFlowLabelKeysLabel,
-		otelcollector.ApertureClassifiersLabel,
-		otelcollector.ApertureClassifierErrorsLabel,
-		otelcollector.ApertureServicesLabel,
-		otelcollector.ApertureControlPointLabel,
-		otelcollector.ApertureResponseStatusLabel,
-	}
-
-	_includeAttributesHTTP = []string{
-		otelcollector.HTTPStatusCodeLabel,
-		otelcollector.HTTPRequestContentLength,
-		otelcollector.HTTPResponseContentLength,
-	}
-
-	_includeAttributesSDK = []string{
-		otelcollector.ApertureFeatureStatusLabel,
-	}
-
-	includeListHTTP = otelcollector.FormIncludeList(append(_includeAttributesCommon, _includeAttributesHTTP...))
-	includeListSDK  = otelcollector.FormIncludeList(append(_includeAttributesCommon, _includeAttributesSDK...))
-)
-
-func enforceIncludeListHTTP(attributes pcommon.Map) {
-	otelcollector.EnforceIncludeList(attributes, includeListHTTP)
-}
-
-func enforceIncludeListSDK(attributes pcommon.Map) {
-	otelcollector.EnforceIncludeList(attributes, includeListSDK)
 }
